requirements.txt
# VocalPoint Dependencies

# Gemini AI
google-genai>=1.0.0

# API Server
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.0.0
